cobraの構成に合わせてメソッドを移動
